split pipeline with base classes
