+++ python/pyspark/ml/tests.py
@@ -37,7 +37,7 @@ else:
 from shutil import rmtree
 import tempfile
 
-from pyspark.ml import Estimator, Model, Pipeline, Transformer
+from pyspark.ml import Estimator, Model, Pipeline, PipelineModel, Transformer
 from pyspark.ml.classification import LogisticRegression
 from pyspark.ml.clustering import KMeans
 from pyspark.ml.evaluation import RegressionEvaluator
@@ -144,6 +144,25 @@ class PipelineTests(PySparkTestCase):
         self.assertEqual(4, model2.dataset_index)
         self.assertEqual(5, transformer3.dataset_index)
         self.assertEqual(6, dataset.index)
+
+        # Test pipeline save/load
+        path = tempfile.mkdtemp()
+        pipeline_path = path + "/pipeline"
+        pipeline.save(pipeline_path)
+        loaded_pipeline = Pipeline.load(pipeline_path)
+        self.assertEqual(pipeline.getStages().size, loaded_pipeline.getStages().size)
+        for p, loaded_p in zip(pipeline.getStages(), loaded_pipeline.getStages()):
+            self.assertEqual(type(p), type(loaded_p))
+            self.assertEqual(p.dataset_index, loaded_p.dataset_index)
+            self.assertEqual(p.getFake(), loaded_p.getFake)
+        model_path = path + "/model"
+        pipeline_model.save(model_path)
+        loaded_model = PipelineModel.load(model_path)
+        self.assertEqual(pipeline_model.stages.size, loaded_pipeline.stages.size)
+        for p, loaded_p in zip(pipeline_model.stages, loaded_model.stages):
+            self.assertEqual(type(p), type(loaded_p))
+            self.assertEqual(p.dataset_index, loaded_p.dataset_index)
+            self.assertEqual(p.getFake(), loaded_p.getFake)
 
 
 class TestParams(HasMaxIter, HasInputCol, HasSeed):
